% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Mgt_Functions.R
\name{dcode}
\alias{dcode}
\title{Convert a factor into a set of dummy codes.}
\usage{
dcode(x, y, stem)
}
\arguments{
\item{x}{A data frame containing the factor varable to be dummy coded.}

\item{y}{The factor to be recoded.}

\item{stem}{A character value for the stem of the variable names for the
resulting dummy codes.}
}
\value{
An updated data frame containing the new dummy-coded variables.
}
\description{
The function converts a factor variable into a set of dummy
  codes then adds them to the data frame on which it operates.
}
\details{
=============================================================================

Each level of the factor will be turned into a new binary variable
  in the data frame x with a name of the form stem_level.
}
