% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Mgt_Functions.R
\name{dfsummary}
\alias{dfsummary}
\title{Summarize properties of criminal history records data frames.}
\usage{
dfsummary(x, label = NULL, dvar)
}
\arguments{
\item{x}{A data frame containing variables named OID and IID.}

\item{label}{An optional vector of row names for the resulting data frame.}

\item{dvar}{A date variable to be summarized. The earliest and latest date
values will be extracted and stored as variables in the result.}
}
\value{
A data frame containing variables N.Records, N.Incidents (based on
  unique values for x$IID), N.Offenders (based on unique values for x$OID),
  Earliest, and Latest.
}
\description{
This function summarizes specific properties of data frames
  into a short data frame.
}
\details{
This function counts the number of records in the data frame, .
}
