% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Mgt_Functions.R
\name{getARRs}
\alias{getARRs}
\title{Get criminal history arrest offense records for an incident.}
\usage{
getARRs(iid, adata)
}
\arguments{
\item{iid}{A single character value for an incident ID (IID).}

\item{adata}{A data frame (or tibble) containing criminal history arrest
offense records.}
}
\value{
A tibble containing the variables AID, ADate, OAgeA, ACount, AClass,
  ACat12, AType, ADisp, ACounty, AMImp.
}
\description{
Retrieves the criminal history arrest offense records associated
  with a specific incident in preparation for subsequent display in a case
  history.
}
\details{
=============================================================================

This function subsets the data frame supplied in the arrests
  parameter down to only the records with IIDs matching the supplied iid
  parameter. It also selects a subset of variables.
}
