% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Mgt_Functions.R
\name{getCHGs}
\alias{getCHGs}
\title{Get criminal history prosecutor charge records for an incident.}
\usage{
getCHGs(iid, cdata)
}
\arguments{
\item{iid}{A single character value for an incident ID (IID).}

\item{cdata}{A data frame (or tibble) containing criminal history
prosecutor charge records.}
}
\value{
A tibble containing the variables CID, CDate, OAgeC, CCount, CCit,
  CCat12, CCounty, CMImp.
}
\description{
Retrieves the criminal history prosecutor charge records
  associated with a specific incident in preparation for subsequent display
  in a case history.
}
\details{
=============================================================================

This function subsets the data frame supplied in the arrests
  parameter down to only the records with IIDs matching the supplied iid
  parameter. It also selects a subset of variables.
}
