% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Mgt_Functions.R
\name{getIIDs}
\alias{getIIDs}
\title{Get all the incident IDs (IIDs) for an offender.}
\usage{
getIIDs(oid, idata)
}
\arguments{
\item{oid}{A single character value for an offender's OID.}

\item{idata}{A data frame (or tibble) containing criminal history incident
records.}
}
\value{
A character vector containing IIDs from all incident records where
  offender OID = oid.
}
\description{
Returns a vector of all the incident IDs (IIDs) associated with
  the specified offender ID (OID).
}
\details{
=============================================================================

This function subsets the data frame supplied in the incidents
  parameter to retain only records with OIDs matching the supplied oid
  parameter. It them extracts the incident IDs from the IID variable and
  returns them.
}
