% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Mgt_Functions.R
\name{getINC}
\alias{getINC}
\title{Get a criminal history incident record.}
\usage{
getINC(iid, incidents)
}
\arguments{
\item{iid}{A single character value for an incident ID (IID).}

\item{incidents}{A data frame (or tibble) containing criminal history
incident records.}
}
\value{
A tibble containing the variables OID, IID, IDate, OAgeI, CSC_ARR,
  CSC_CHG, CSC_JUD, CSC_CON, CSC_ANY, and CSC_ACC.
}
\description{
Retrieves a criminal history incident record in preparation
  for subsequent display in a case history.
}
\details{
=============================================================================

This function subsets the data frame supplied in the incidents
  parameter down to only the record with an IID matching the supplied iid
  parameter. It also selects a subset of variables. Along the way it renames
  HXCat12_5 to CSC_ACC. That binary history variable represents whether the
  offender was arrested, charged, or convicted of criminal sexual conduct in
  connection with the incident, which is related to but slightly narrower
  than CSC_ANY, which is a binary variable indicating whether the offender
  was arrested, charged, or adjudicated for any criminal sexual conduct
  charges (regardless of judicial disposition) in connection with the
  incident.
}
