% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Mgt_Functions.R
\name{getJUDs}
\alias{getJUDs}
\title{Get criminal history judicial charge records for an incident.}
\usage{
getJUDs(iid, jcharges)
}
\arguments{
\item{iid}{A single character value for an incident ID (IID).}

\item{jcharges}{A data frame (or tibble) containing criminal history
judicial charge records.}
}
\value{
A tibble containing the variables JID, JDate, OAgeJ, JCount, JCit,
  JCat12, JCounty, JDateImp.
}
\description{
Retrieves the criminal history judicial charge records
  associated with a specific incident in preparation for subsequent display
  in a case history.
}
\details{
=============================================================================

This function subsets the data frame supplied in the arrests
  parameter down to only the records with IIDs matching the supplied iid
  parameter. It also selects a subset of variables.
}
