% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice_Functions.R
\name{panel.mybox}
\alias{panel.mybox}
\title{Panel function for overlaying a box plot on a strip plot.}
\usage{
panel.mybox(...)
}
\arguments{
\item{...}{Arguments are passed to panel.stripplot() and panel.bwplot().}
}
\description{
This function allows you to overlay a boxplot on top of a
  jittered strip plot.
}
\details{
This custom panel function designed for use with the
  lattice::bwplot() function simplifies overlaying a box plot on top of a
  jittered strip plot.
}
\seealso{
\code{\link[lattice]{panel.stripplot}} and
  \code{\link[lattice]{panel.bwplot}}.
}
