% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lattice_Functions.R
\name{panel.mydensity}
\alias{panel.mydensity}
\title{Panel function for overlaying a normal curve on a density plot.}
\usage{
panel.mydensity(x, ...)
}
\arguments{
\item{x}{Data points for which density is to be estimated, as discussed in
help for panel.densityplot().}

\item{...}{Arguments are passed to panel.densityplot() and panel.mathdensity().}
}
\description{
This function allows you to overlay a normal distribution curve
  on top of a density plot.
}
\details{
=============================================================================

This panel function designed for use with lattice::densityplot()
  simplifies overlaying a normal distribution curve on top of a density plot.
}
\seealso{
\code{\link[lattice]{densityplot}},
  \code{\link[lattice]{panel.densityplot}}, and
  \code{\link[lattice]{panel.mathdensity}}.
}
