% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Mgt_Functions.R
\name{rvlabel}
\alias{rvlabel}
\title{Replace a specific value label in an labelled_spss object.}
\usage{
rvlabel(x, from = "Excluded (user-missing)", to = "Excluded user-missing")
}
\arguments{
\item{x}{A variable with class = labelled_spss.}

\item{from}{A character value for the old label to be replaced.}

\item{to}{A character value for the new label that will replace the old one.}
}
\value{
An updated variable with class = labelled_spss.
}
\description{
The function replaces a specific value label with a new label.
}
\details{
=============================================================================

This function is primarily used to get around the fact that
  parentheses in the value labels imported with some SPSS data files cause
  a problem with some of our R code. this allows us to quickly replace those
  value labels.
}
