% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSACHR-package.R
\docType{package}
\name{SSACHR-package}
\alias{SSACHR}
\alias{SSACHR-package}
\title{SSACHR: Serial Sexual Assault Study Criminal History Records Paper Research 
  Compendium}
\description{
This package and corresponding GitHub repository are intended to
  enhance the reproducibility of a research paper by serving as a research
  compendium (Marwick, Boettiger, & Mullen, 2018). The principal investigator 
  for the study is Dr. Rebecca Campbell (Professor, Department of Psychology, 
  Michigan State University). The paper will present descriptive data about
  the criminal histories of suspected serial sexual offenders using data from
  both criminal history records and forensic DNA testing of sexual assault 
  kits (SAKs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sjpierce/SSACHR}
}

}
\author{
\strong{Maintainer}: Steven J. Pierce \email{pierces1@msu.edu} (\href{https://orcid.org/0000-0002-0679-3019}{ORCID}) [copyright holder]

}
\keyword{internal}
